Renamed OAuth2 /token endpoint handlers to ITokenRequestHandlers
This makes it a little more intuitive given their purpose in the system.
