+++ kangaroo-server-authz/src/main/java/net/krotscheck/kangaroo/authz/oauth2/OAuthAPI.java
@@ -25,10 +25,10 @@ import net.krotscheck.kangaroo.authz.oauth2.factory.CredentialsFactory;
 import net.krotscheck.kangaroo.authz.oauth2.filter.ClientAuthorizationFilter;
 import net.krotscheck.kangaroo.authz.oauth2.resource.AuthorizationService;
 import net.krotscheck.kangaroo.authz.oauth2.resource.TokenService;
-import net.krotscheck.kangaroo.authz.oauth2.resource.grant.AuthorizationCodeGrantHandler;
-import net.krotscheck.kangaroo.authz.oauth2.resource.grant.ClientCredentialsGrantHandler;
-import net.krotscheck.kangaroo.authz.oauth2.resource.grant.OwnerCredentialsGrantHandler;
-import net.krotscheck.kangaroo.authz.oauth2.resource.grant.RefreshTokenGrantHandler;
+import net.krotscheck.kangaroo.authz.oauth2.resource.token.AuthorizationCodeGrantHandler;
+import net.krotscheck.kangaroo.authz.oauth2.resource.token.ClientCredentialsGrantHandler;
+import net.krotscheck.kangaroo.authz.oauth2.resource.token.OwnerCredentialsGrantHandler;
+import net.krotscheck.kangaroo.authz.oauth2.resource.token.RefreshTokenGrantHandler;
 import net.krotscheck.kangaroo.authz.oauth2.tasks.TokenCleanupTask;
 import net.krotscheck.kangaroo.common.config.ConfigurationFeature;
 import net.krotscheck.kangaroo.common.exception.ExceptionFeature;
