+++ kangaroo-server-authz/src/main/java/net/krotscheck/kangaroo/authz/oauth2/resource/token/RefreshTokenGrantHandler.java
@@ -15,14 +15,15 @@
  * limitations under the License.
  */
 
-package net.krotscheck.kangaroo.authz.oauth2.resource.grant;
+package net.krotscheck.kangaroo.authz.oauth2.resource.token;
 
+import net.krotscheck.kangaroo.authz.common.database.entity.ApplicationScope;
 import net.krotscheck.kangaroo.authz.common.database.entity.Client;
 import net.krotscheck.kangaroo.authz.common.database.entity.ClientType;
 import net.krotscheck.kangaroo.authz.common.database.entity.OAuthToken;
 import net.krotscheck.kangaroo.authz.common.database.entity.OAuthTokenType;
+import net.krotscheck.kangaroo.authz.common.util.ValidationUtil;
 import net.krotscheck.kangaroo.authz.oauth2.exception.RFC6749.InvalidGrantException;
-import net.krotscheck.kangaroo.authz.oauth2.exception.RFC6749.InvalidRequestException;
 import net.krotscheck.kangaroo.authz.oauth2.resource.TokenResponseEntity;
 import org.glassfish.hk2.utilities.binding.AbstractBinder;
 import org.glassfish.jersey.process.internal.RequestScoped;
@@ -30,19 +31,17 @@ import org.hibernate.Session;
 
 import javax.inject.Inject;
 import javax.ws.rs.core.MultivaluedMap;
-import javax.ws.rs.core.UriBuilder;
-import java.net.URI;
-import java.util.List;
+import java.util.SortedMap;
 import java.util.UUID;
 
 /**
- * This grant type handler takes care of the "authorization_code" grant_type
- * OAuth flow. Given an authorization code provided via the auth code flow,
- * this handler will validate the request and issue a token.
+ * This token type handler takes care of the "refresh_token" grant_type
+ * OAuth flow. Its job is to handle all refresh token requests, by
+ * invalidating any previous tokens and issuing new ones.
  *
  * @author Michael Krotscheck
  */
-public final class AuthorizationCodeGrantHandler implements IGrantTypeHandler {
+public final class RefreshTokenGrantHandler implements ITokenRequestHandler {
 
     /**
      * Hibernate session, injected.
@@ -50,104 +49,87 @@ public final class AuthorizationCodeGrantHandler implements IGrantTypeHandler {
     private final Session session;
 
     /**
-     * Create a new instance of this grant handler.
+     * Create a new instance of this token handler.
      *
      * @param session Injected hibernate session.
      */
     @Inject
-    public AuthorizationCodeGrantHandler(final Session session) {
+    public RefreshTokenGrantHandler(final Session session) {
         this.session = session;
     }
 
     /**
-     * Apply the client credentials flow to this request.
+     * Handle a specific token type request.
      *
-     * @param client   The Client to use.
-     * @param formData Raw form data for the request.
-     * @return A response indicating the result of the request.
+     * @param client   The client.
+     * @param formData Additional form data.
+     * @return A token response entity with the new token.
      */
     @Override
     public TokenResponseEntity handle(final Client client,
                                       final MultivaluedMap<String, String>
                                               formData) {
         // Make sure the client is the correct type.
-        if (!client.getType().equals(ClientType.AuthorizationGrant)) {
+        ClientType type = client.getType();
+        if (!type.equals(ClientType.OwnerCredentials)
+                && !type.equals(ClientType.AuthorizationGrant)) {
             throw new InvalidGrantException();
         }
 
-        UUID authCodeId;
-        URI redirect;
+        // Cast the token to a UUID.
+        UUID refreshId;
         try {
-            authCodeId = UUID.fromString(getOne(formData, "code"));
-            redirect = UriBuilder.fromUri(getOne(formData, "redirect_uri"))
-                    .build();
+            refreshId = UUID.fromString(formData.getFirst("refresh_token"));
         } catch (IllegalArgumentException | NullPointerException e) {
             throw new InvalidGrantException();
         }
 
-        // Retrieve the authorization code
-        OAuthToken authCode = session.get(OAuthToken.class, authCodeId);
-        if (authCode == null) {
+        // Attempt to get the refresh token
+        OAuthToken refreshToken = session.get(OAuthToken.class, refreshId);
+        if (refreshToken == null || !refreshToken.getTokenType()
+                .equals(OAuthTokenType.Refresh)) {
             throw new InvalidGrantException();
         }
 
-        // Check to see if it's expired.
-        if (authCode.isExpired()) {
+        // Make sure the refresh token is not expired.
+        if (refreshToken.isExpired()) {
             throw new InvalidGrantException();
         }
 
-        // Make sure that the client and auth code match
-        if (!authCode.getClient().equals(client)) {
-            throw new InvalidGrantException();
-        }
-
-        // Ensure that the redirect URL matches.
-        if (!redirect.equals(authCode.getRedirect())) {
-            throw new InvalidGrantException();
-        }
+        // Make sure the requested scopes are valid for the refresh token.
+        SortedMap<String, ApplicationScope> requestedScopes =
+                ValidationUtil.revalidateScope(
+                        formData.getFirst("scope"),
+                        refreshToken.getScopes(),
+                        refreshToken.getIdentity().getUser().getRole()
+                );
 
         // Ensure that we retrieve a state, if it exists.
         String state = formData.getFirst("state");
 
-        // Go ahead and create the token.
+        // Go ahead and create the tokens.
         OAuthToken newAuthToken = new OAuthToken();
         newAuthToken.setClient(client);
         newAuthToken.setTokenType(OAuthTokenType.Bearer);
         newAuthToken.setExpiresIn(client.getAccessTokenExpireIn());
-        newAuthToken.setScopes(authCode.getScopes());
+        newAuthToken.setScopes(requestedScopes);
 
         OAuthToken newRefreshToken = new OAuthToken();
         newRefreshToken.setClient(client);
         newRefreshToken.setTokenType(OAuthTokenType.Refresh);
         newRefreshToken.setExpiresIn(client.getRefreshTokenExpireIn());
-        newRefreshToken.setScopes(authCode.getScopes());
+        newRefreshToken.setScopes(requestedScopes);
         newRefreshToken.setAuthToken(newAuthToken);
 
         session.save(newAuthToken);
         session.save(newRefreshToken);
-        session.delete(authCode);
+        if (refreshToken.getAuthToken() != null) {
+            session.delete(refreshToken.getAuthToken());
+        }
+        session.delete(refreshToken);
 
         return TokenResponseEntity.factory(newAuthToken, newRefreshToken,
                 state);
-    }
-
-    /**
-     * Helper method, extracts one (and only one) value from a multivaluedmap.
-     *
-     * @param values The map of values.
-     * @param key    The key to get.
-     * @return The value retrieved, but only if only one exists for this key.
-     */
-    private String getOne(final MultivaluedMap<String, String> values,
-                          final String key) {
-        List<String> listValues = values.get(key);
-        if (listValues == null) {
-            throw new InvalidRequestException();
-        }
-        if (listValues.size() != 1) {
-            throw new InvalidRequestException();
-        }
-        return listValues.get(0);
     }
 
     /**
@@ -157,9 +139,9 @@ public final class AuthorizationCodeGrantHandler implements IGrantTypeHandler {
 
         @Override
         protected void configure() {
-            bind(AuthorizationCodeGrantHandler.class)
-                    .to(IGrantTypeHandler.class)
-                    .named("authorization_code")
+            bind(RefreshTokenGrantHandler.class)
+                    .to(ITokenRequestHandler.class)
+                    .named("refresh_token")
                     .in(RequestScoped.class);
         }
     }
