+++ kangaroo-server-authz/src/test/java/net/krotscheck/kangaroo/authz/admin/v1/resource/AbstractServiceBrowseTest.java
@@ -60,7 +60,7 @@ public abstract class AbstractServiceBrowseTest<T extends AbstractAuthzEntity>
         extends AbstractResourceTest {
 
     /**
-     * The scope to grant the issued token.
+     * The scope to token the issued token.
      */
     private final String tokenScope;
