+++ kangaroo-server-authz/src/test/java/net/krotscheck/kangaroo/authz/oauth2/rfc6749/Section410AuthorizationCodeGrantTest.java
@@ -1035,7 +1035,7 @@ public final class Section410AuthorizationCodeGrantTest
     }
 
     /**
-     * Assert that missing a grant type errors.
+     * Assert that missing a token type errors.
      */
     @Test
     public void testTokenNoGrant() {
@@ -1301,7 +1301,7 @@ public final class Section410AuthorizationCodeGrantTest
     }
 
     /**
-     * Assert that an invalid grant type errors.
+     * Assert that an invalid token type errors.
      */
     @Test
     public void testTokenInvalidGrantTypePassword() {
@@ -1331,7 +1331,7 @@ public final class Section410AuthorizationCodeGrantTest
     }
 
     /**
-     * Assert that an invalid grant type errors.
+     * Assert that an invalid token type errors.
      */
     @Test
     public void testTokenInvalidGrantTypeRefreshToken() {
@@ -1361,7 +1361,7 @@ public final class Section410AuthorizationCodeGrantTest
     }
 
     /**
-     * Assert that an invalid grant type errors.
+     * Assert that an invalid token type errors.
      */
     @Test
     public void testTokenInvalidGrantTypeClientCredentials() {
@@ -1391,7 +1391,7 @@ public final class Section410AuthorizationCodeGrantTest
     }
 
     /**
-     * Assert that an unknown grant type errors.
+     * Assert that an unknown token type errors.
      */
     @Test
     public void testTokenUnknownGrantType() {
