+++ kangaroo-server-authz/src/test/java/net/krotscheck/kangaroo/authz/oauth2/rfc6749/Section430OwnerPasswordTest.java
@@ -192,7 +192,7 @@ public final class Section430OwnerPasswordTest
     }
 
     /**
-     * Assert that missing a grant type errors.
+     * Assert that missing a token type errors.
      */
     @Test
     public void testTokenNoGrant() {
@@ -405,7 +405,7 @@ public final class Section430OwnerPasswordTest
     }
 
     /**
-     * Assert that an invalid grant type errors.
+     * Assert that an invalid token type errors.
      */
     @Test
     public void testTokenInvalidGrantTypeRefreshToken() {
@@ -430,7 +430,7 @@ public final class Section430OwnerPasswordTest
     }
 
     /**
-     * Assert that an invalid grant type errors.
+     * Assert that an invalid token type errors.
      */
     @Test
     public void testTokenInvalidGrantTypeClientCredentials() {
@@ -455,7 +455,7 @@ public final class Section430OwnerPasswordTest
     }
 
     /**
-     * Assert that an unknown grant type errors.
+     * Assert that an unknown token type errors.
      */
     @Test
     public void testTokenUnknownGrantType() {
